Completed preliminary reimplementation of type()
Updated test unit to match.

Throw an exception instead of returning None/False in some
functions.

+++ src/util.h
@@ -28,4 +28,4 @@
 int checksum(const u8 * buf, size_t len);
 void *mem_chunk(size_t base, size_t len, const char *devmem);
 int write_dump(size_t base, size_t len, const void *data, const char *dumpfile, int add);
-int is_int(const char *s);
+long is_int(const char *s);
